Change u128 to i32

+++ src/main.rs
@@ -4,11 +4,11 @@ mod math;
 mod user_input;
 
 fn main() {
-    let user_input: u128 = user_input::get_user_input("Enter a positive integer: ");
+    let user_input: i32 = user_input::get_user_input("Enter a positive integer: ");
 
-    let mut counter: u128 = 0;
-    let mut prime_counter: u128 = 0;
-    let mut primes: Vec<u128> = vec![];
+    let mut counter: i32 = 0;
+    let mut prime_counter: i32 = 0;
+    let mut primes: Vec<i32> = vec![];
 
     while prime_counter <= user_input {
         let is_prime: bool = math::is_prime_number(counter);
